+++ lib/appconfig.php
@@ -278,6 +278,13 @@ class AppConfig {
     public $_customization_macros = "customization_macros";
 
     /**
+     * The config key for the plugins
+     *
+     * @var string
+     */
+    public $_customizationPlugins = "customization_plugins";
+
+    /**
      * @param string $AppName - application name
      */
     public function __construct($AppName) {
@@ -867,6 +874,26 @@ class AppConfig {
      */
     public function GetCustomizationMacros() {
         return $this->config->getAppValue($this->appName, $this->_customization_macros, "true") === "true";
+    }
+
+    /**
+     * Save plugins setting
+     *
+     * @param bool $value - enable macros
+     */
+    public function SetCustomizationPlugins($value) {
+        $this->logger->info("Set plugins enabled: " . json_encode($value), ["app" => $this->appName]);
+
+        $this->config->setAppValue($this->appName, $this->_customizationPlugins, json_encode($value));
+    }
+
+    /**
+     * Get plugins setting
+     *
+     * @return bool
+     */
+    public function GetCustomizationPlugins() {
+        return $this->config->getAppValue($this->appName, $this->_customizationPlugins, "true") === "true";
     }
 
     /**