+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Added
 - Dutch translation
 - Chinese (Traditional, Taiwan), Basque (Spain), Armenian and Malay (Malaysia) empty file templates
+- disable plugins setting
 
 ## Changed
 - fix edit locked file